test/unit/Raffle: check that we are not allowed to enter in the critical section
enterRaffle should disallow entering the lottery, if the lottery is set to
closed/calculating phase.

We simulate this by artifially progressing the time.
Mine a new block, then call performUpkeep, which should
put the contract in CALCULATING state.

Then we try to enter the contract with another account.
And it should be reverted with the custom error, Raffle__NotOpen

NOTE: since we are testing on a development chain, we need to add
our consumer contract to the subscription ourselves.
else it will fail with InvalidConsumer error.

Solved thanks to a random helpful guy on the internet:
https://github.com/smartcontractkit/full-blockchain-solidity-course-js/discussions/1375#discussioncomment-3333258

Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>